+++ packages/@runlightyear/lightyear/src/base/auth.ts
@@ -58,7 +58,7 @@ export type AuthData = {
   /**
    * Store app-specific data beyond the OAuth spec
    */
-  extraData?: string | null;
+  extraData?: Record<string, any> | null;
   /**
    * Custom app data
    */