+++ packages/@runlightyear/node/src/Lightyear.ts
@@ -2,6 +2,7 @@ import { ModelSynchronizer } from "./ModelSynchronizer";
 
 export interface LightyearProps {
   apiKey: string;
+  baseUrl?: string;
 }
 
 export interface SyncConfig {
@@ -11,20 +12,22 @@ export interface SyncConfig {
   };
 }
 
-const PREFIX = "http://localhost:3000";
+const DEFAULT_BASE_URL = "http://localhost:3000";
 
 export class Lightyear {
   apiKey: string;
+  baseUrl: string;
 
   constructor(props: LightyearProps) {
     this.apiKey = props.apiKey;
+    this.baseUrl = props.baseUrl ?? DEFAULT_BASE_URL;
   }
 
   async getModels(props: { collection: string }) {
     const { collection } = props;
 
     const response = await fetch(
-      `${PREFIX}/api/v1/envs/dev/collections/${collection}/models`,
+      `${this.baseUrl}/api/v1/envs/dev/collections/${collection}/models`,
       {
         headers: {
           "Content-Type": "application/json",
@@ -51,7 +54,7 @@ export class Lightyear {
     const { collection, model, userId, objectId, updatedAt, data } = props;
 
     const response = await fetch(
-      `${PREFIX}/api/v1/envs/dev/collections/${collection}/models/${model}/objects/upsert`,
+      `${this.baseUrl}/api/v1/envs/dev/collections/${collection}/models/${model}/objects/upsert`,
       {
         method: "POST",
         headers: {
@@ -83,7 +86,7 @@ export class Lightyear {
     const { collection, model, objectId } = props;
 
     const response = await fetch(
-      `${PREFIX}/api/v1/envs/dev/collections/${collection}/models/${model}/objects/delete`,
+      `${this.baseUrl}/api/v1/envs/dev/collections/${collection}/models/${model}/objects/delete`,
       {
         method: "POST",
         headers: {
@@ -104,7 +107,7 @@ export class Lightyear {
     const { collection, model } = props;
 
     const response = await fetch(
-      `${PREFIX}/api/v1/envs/dev/collections/${collection}/models/${model}/objects/delta`,
+      `${this.baseUrl}/api/v1/envs/dev/collections/${collection}/models/${model}/objects/delta`,
       {
         method: "POST",
         headers: {